docker/transaction-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
prometheus-client==0.19.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
